Extract two server functions: `setupServer` and `shutdownServer`, make `startServer` use them to accept connections
